WIP! Code move
Split up remote build steps into git, api and the build itself.
